move RequestSender and its subclass to new package(protocol)
