+++ src/test/scala/com/naokia/groonga4s/response/LoadResponseParserSpec.scala
@@ -3,11 +3,7 @@ package com.naokia.groonga4s.response
 import org.specs2.mutable.Specification
 import com.fasterxml.jackson.databind._
 
-/**
- * Created by naoki on 15/10/30.
- */
-class LoadResponseParserTest extends Specification {
-
+class LoadResponseParserSpec extends Specification {
   "LoadResponseParserTest" should {
     "parse" in {
       val jsonStr = """[[0,1446206383.85375,0.0631723403930664],1]"""
